feat(libyml): :sparkles: Create a new file called internal.rs for internal helper functions
