+++ src/emitter.rs
@@ -1,5 +1,5 @@
-use crate::api::{yaml_queue_extend, yaml_stack_extend};
 use crate::externs::{strcmp, strlen, strncmp};
+use crate::internal::{yaml_queue_extend, yaml_stack_extend};
 use crate::memory::{yaml_free, yaml_strdup};
 use crate::ops::{ForceAdd as _, ForceMul as _};
 use crate::success::{Success, FAIL, OK};
