+++ src/loader.rs
@@ -1,5 +1,5 @@
-use crate::api::yaml_stack_extend;
 use crate::externs::{memset, strcmp};
+use crate::internal::yaml_stack_extend;
 use crate::memory::{yaml_free, yaml_malloc, yaml_strdup};
 use crate::success::{Success, FAIL, OK};
 use crate::yaml::yaml_char_t;
