+++ src/parser.rs
@@ -1,5 +1,5 @@
-use crate::api::yaml_stack_extend;
 use crate::externs::{memcpy, memset, strcmp, strlen};
+use crate::internal::yaml_stack_extend;
 use crate::memory::{yaml_free, yaml_malloc, yaml_strdup};
 use crate::ops::ForceAdd as _;
 use crate::scanner::yaml_parser_fetch_more_tokens;
